RPi/requirements.txt
click==8.1.7
Flask==3.0.3
pyserial==3.5
